# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InmoDB CRM API",
        "description": "API Documentation for InmoDB Real Estate & CRM Platform",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/crm/deals": {
            "get": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "List User's Deals",
                "operationId": "a0141aeb647e5693148d3701e1995e88",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pipeline_id",
                        "in": "query",
                        "description": "Filter by Pipeline ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 500
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Big Deal 2024"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 50000
                                                    },
                                                    "stage_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "open"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Create New Deal",
                "operationId": "28faada542dc3d4ffb7ceb0b7d133b1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "pipeline_id",
                                    "stage_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Opportunity"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150000
                                    },
                                    "pipeline_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stage_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "won",
                                            "lost"
                                        ],
                                        "example": "open"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deal created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crm/deals/{id}": {
            "get": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Get Deal Details",
                "operationId": "17ce7c2969360558232ba06253f442d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Deal ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "attributes": {
                                                    "$ref": "#/components/schemas/Deal"
                                                },
                                                "timeline": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "associations": {
                                                    "properties": {
                                                        "contacts": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "companies": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "tickets": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "properties": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deal not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Update Deal",
                "operationId": "1fb05cbf18d38346b29fb1311e65dcea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Delete Deal",
                "operationId": "d0336a424ab066f2d2b8bc86af6cb857",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crm/deals/{id}/stage": {
            "post": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Move Deal Stage",
                "operationId": "896aa5111c8b1df7f11f62114583456e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stage_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stage updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crm/deals/{id}/win": {
            "post": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Mark Deal as Won",
                "operationId": "0908576639545ec0fc8dd39e296a3825",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deal Won"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crm/deals/{id}/lose": {
            "post": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Mark Deal as Lost",
                "operationId": "e3faf1b2fd2c22130a687bdd12f8bf73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lost_reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deal Lost"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/crm/deals/{id}/analytics": {
            "get": {
                "tags": [
                    "CRM - Deals"
                ],
                "summary": "Get Deal Analytics",
                "operationId": "efea02c35d49c18e825456c30c6b5634",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analytics data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Deal": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "pipeline_id": {
                        "type": "integer"
                    },
                    "stage_id": {
                        "type": "integer"
                    },
                    "expected_close_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}